requirements/common.txt
# Common requirements
coreapi==2.3.3
cryptography==41.0.1
Django==4.2.2
django-bootstrap5==23.1
django-filter==23.2
djangorestframework==3.14
docutils==0.20.1
drf-yasg==1.21.5
GitPython==3.1.31
lxml==4.9.2
mysqlclient==2.1.1
python-dateutil==2.8.2